retry (#18)
* modify udo

modify udo

* misslie modiy

* cluster final!

cluster bulley and cluster move add.

* temprory camera Mgr adding

adding persent 20%

* deleate CameraActionManager

Modfiy..

* filters_Test

* Update FieldCameraZoomInState.h

* camera state

state

* Modfiy Include, Soruce

 camerastate

* cluster distance dead

filed  camera / ing

* camera 10 %

* camera

+++ Client/Source/Player_ClusterMissile.cpp
@@ -16,16 +16,29 @@ Player_ClusterMissile::~Player_ClusterMissile()
 
 void Player_ClusterMissile::ItemShot()
 {
+	m_detector = Core::GetInstance()->FindObjectByName(OBJECT_TAG_UI, L"detector")->GetComponent<EnemyDetector>();
 
-	SphereCollider::Desc cluster_col;
-	cluster_col.collisionLayer = COLLISION_LAYER_BULLET_PLAYER;
-	m_bullet->AddComponent<SphereCollider>(&cluster_col);
+	if (m_detector == nullptr)
+		return;
 
-	//m_bullet->GetTransform()->position = m_bullet->GetTransform()->GetWorldPosition();
+	auto lockonTarget = m_detector->GetLockOnTarget();
+
+	if (lockonTarget == nullptr)
+		return;
+
+
+
+
+	m_bullet->GetTransform()->position = m_bullet->GetTransform()->GetWorldPosition();
 	m_bullet->GetTransform()->DeleteParent();
 
 	ClusterBulletMove::Desc bulletinfo;
 	bulletinfo.speed = m_itemInfo.weaponSpeed;
+	Vector2 screenPos = Core::GetInstance()->GetMainCamera()->WorldToScreenPos(lockonTarget->GetTransform()->position);
+
+	bulletinfo.screenPos = Vector2(screenPos.x * cosf(D3DXToRadian(90.f)), screenPos.y * sinf(D3DXToRadian(90.f)));
+	bulletinfo.target = lockonTarget;
+	bulletinfo.accAngle = bulletinfo.speed;
 
 	m_bullet->AddComponent<ClusterBulletMove>(&bulletinfo);
